+++ backend/tests/__init__.py
@@ -1,0 +1,3 @@
+import unittest
+
+unittest.TestLoader.sortTestMethodsUsing = None
